+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changes
 - Change display of true and false values in the `info` and `search` commands.
+- Prebuilds now have a toml metadata file instead of the checksum file, containing the checksum and size of the prebuild.
 
 
 ## [v0.0.3](https://github.com/pack-it/packit/compare/0.0.2...0.0.3) - 2026-07-11